+++ include/UnivariatePolynomial.h
@@ -95,12 +95,12 @@ public:
   {
     std::vector<K> new_a(std::max(a.size(), p.coefficient().size()), 0);
 
-    for (int a_i = 0; a_i < a.size(); a_i++)
+    for (size_t a_i = 0; a_i < a.size(); a_i++)
     {
       new_a[a_i] = a[a_i];
     }
 
-    for (int p_a_i = 0; p_a_i < p.coefficient().size(); p_a_i++)
+    for (size_t p_a_i = 0; p_a_i < p.coefficient().size(); p_a_i++)
     {
       new_a[p_a_i] += p.coefficient()[p_a_i];
     }
@@ -123,9 +123,9 @@ public:
     }
 
     std::vector<K> new_a(this->degree() + p.degree() + 1, 0);
-    for (int a_i = 0; a_i < a.size(); a_i++)
+    for (size_t a_i = 0; a_i < a.size(); a_i++)
     {
-      for (int p_a_i = 0; p_a_i < p.coefficient().size(); p_a_i++)
+      for (size_t p_a_i = 0; p_a_i < p.coefficient().size(); p_a_i++)
       {
         new_a[a_i + p_a_i] += a[a_i] * p.coefficient()[p_a_i];
       }
@@ -165,7 +165,7 @@ public:
     if (p.degree() != q.degree())
       return false;
 
-    for (int i = 0; i < p.a.size(); i++)
+    for (size_t i = 0; i < p.a.size(); i++)
     {
       if (p.a.at(i) != q.a.at(i))
         return false;
@@ -178,7 +178,7 @@ public:
   {
     os << "[";
 
-    for (int a_i = 0; a_i < p.a.size(); a_i++)
+    for (size_t a_i = 0; a_i < p.a.size(); a_i++)
     {
       os << p.a.at(a_i);
 
@@ -249,7 +249,7 @@ public:
   UnivariatePolynomial differential() const
   {
     auto new_a(a);
-    for (int a_i = 0; a_i < a.size(); a_i++)
+    for (size_t a_i = 0; a_i < a.size(); a_i++)
     {
       new_a[a_i] *= a_i;
     }
